requirements.txt
discord.py==2.1.1
discloud==0.1.0